@@ -1,10 +1,10 @@
 package com.gelostech.pocketbartender.adapters;
 
 import android.content.Context;
-import android.content.Intent;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.core.content.ContextCompat;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -14,7 +14,6 @@ import android.widget.TextView;
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.RequestOptions;
 import com.gelostech.pocketbartender.R;
-import com.gelostech.pocketbartender.activities.CocktailActivity;
 import com.gelostech.pocketbartender.models.HomeModel;
 
 import java.util.List;
