@@ -7,11 +7,11 @@ import android.content.Intent
 import android.net.Uri
 import android.os.Build
 import android.os.Bundle
-import android.support.annotation.RequiresApi
-import android.support.v4.app.Fragment
-import android.support.v4.content.ContextCompat
-import android.support.v7.preference.Preference
-import android.support.v7.preference.PreferenceFragmentCompat
+import androidx.annotation.RequiresApi
+import androidx.fragment.app.Fragment
+import androidx.core.content.ContextCompat
+import androidx.preference.Preference
+import androidx.preference.PreferenceFragmentCompat
 import com.gelostech.pocketbartender.R
 import com.github.javiersantos.materialstyleddialogs.MaterialStyledDialog
 import com.github.javiersantos.materialstyleddialogs.enums.Style
@@ -30,6 +30,7 @@ class PreferencesFragment : PreferenceFragmentCompat(){
         addPreferencesFromResource(R.xml.app_preferences)
 
         aboutApp()
+        more()
         sendEmail()
         rateApp()
         shareApp()
@@ -50,6 +51,24 @@ class PreferencesFragment : PreferenceFragmentCompat(){
                     .setPositiveText("DISMISS")
                     .onPositive { dialog, _ -> dialog.dismiss() }
                     .show()
+
+            true
+        }
+    }
+
+    private fun more() {
+        val more = findPreference("more")
+        more.onPreferenceClickListener = Preference.OnPreferenceClickListener {
+            val uri = Uri.parse(resources.getString(R.string.developer_id))
+            val devAccount = Intent(Intent.ACTION_VIEW, uri)
+
+            devAccount.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY or Intent.FLAG_ACTIVITY_NEW_DOCUMENT or Intent.FLAG_ACTIVITY_MULTIPLE_TASK)
+            try {
+                startActivity(devAccount)
+            } catch (e: ActivityNotFoundException) {
+                startActivity(Intent(Intent.ACTION_VIEW,
+                        Uri.parse(resources.getString(R.string.developer_id))))
+            }
 
             true
         }
