@@ -4,8 +4,8 @@ import android.content.Context
 import android.graphics.Canvas
 import android.graphics.Rect
 import android.graphics.drawable.Drawable
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.RecyclerView
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 
 import com.gelostech.pocketbartender.R
@@ -15,7 +15,7 @@ class RecyclerFormatter {
     class SimpleDividerItemDecoration(context: Context) : RecyclerView.ItemDecoration() {
         private val mDivider: Drawable? = ContextCompat.getDrawable(context, R.drawable.recycler_divider)
 
-        override fun onDrawOver(c: Canvas, parent: RecyclerView, state: RecyclerView.State?) {
+        override fun onDrawOver(c: Canvas, parent: RecyclerView, state: RecyclerView.State) {
             val left = parent.paddingLeft
             val right = parent.width - parent.paddingRight
 
@@ -38,7 +38,7 @@ class RecyclerFormatter {
 
         constructor(context: Context, space: Int) : this(context.resources.getDimensionPixelSize(space)) {}
 
-        override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State?) {
+        override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
             super.getItemOffsets(outRect, view, parent, state)
 
             outRect.set(space, space, space, space)
