@@ -2,9 +2,9 @@ package com.gelostech.pocketbartender.utils
 
 import android.content.Context
 import android.graphics.Typeface
-import android.support.annotation.LayoutRes
-import android.support.design.widget.Snackbar
-import android.support.v4.content.ContextCompat
+import androidx.annotation.LayoutRes
+import com.google.android.material.snackbar.Snackbar
+import androidx.core.content.ContextCompat
 import android.util.AttributeSet
 import android.view.LayoutInflater
 import android.view.View
